Principalmente sono stati aggiunti i meccanismi, sono stati ricontrollati i prefab vecchi, ed è stata aggiunta la UI e il menù
**abbiamo avuto problemi con github**

+ Aggiunto il 3° power up: quello che aumenta la mobilità del player
+ Aggiunto il tappetto salterino
+ Aggiunto il dash
+ Aggiunto il parking gate col pulsante
+ Gestione della morte del player
* Ricontrollati in generale i prefab dei nemici, player e delle trappole, cambiato qualche valore
* Refactoring della test scene, ora è il tutorial

+ Aggiunto il menù e la UI fatti da @massi-fa

+++ Assets/Scripts/Support/Waiter.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 
 public class Waiter : MonoBehaviour
 {
-    public static IEnumerator Active(int nSeconds)
+    public static IEnumerator Active(float nSeconds)
     {
         yield return new WaitForSeconds(nSeconds);
     }
